+++ default-configs/pci.mak
@@ -48,3 +48,4 @@ CONFIG_VIRTIO_INPUT_PCI=y
 CONFIG_VFIO_PCI_QUIRKS=y
 CONFIG_VFIO_PCI_QUIRKS_IGD=y
 CONFIG_VFIO_PCI_QUIRKS_ATI=y
+CONFIG_VFIO_PCI_QUIRKS_NVIDIA=y
