+++ default-configs/x86_64_virt-softmmu.mak
@@ -37,3 +37,4 @@ CONFIG_ACPI_PCI_HOTPLUG=y
 
 CONFIG_PFLASH_CFI01=y
 CONFIG_HARD_DISK=y
+CONFIG_VFIO_PCI_QUIRKS_NVIDIA=y
