+++ hw/vfio/quirks/Makefile.objs
@@ -11,3 +11,9 @@ obj-y += pci-ati.o
 else
 obj-y += pci-ati-stub.o
 endif
+
+ifeq ($(CONFIG_VFIO_PCI_QUIRKS_NVIDIA), y)
+obj-y += pci-nvidia.o
+else
+obj-y += pci-nvidia-stub.o
+endif
